pyproject.toml
[tool.poetry]
name = "website-builder-llm"
version = "0.0.1"
description = "Uses langchain to build a website"
authors = ["Gordy"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
wikipedia = "^1.4.0"
gradio = "^3.38.0"
duckduckgo-search = "^3.8.4"
openai = "^0.27.8"
pygithub = "^1.59.0"
python-dotenv = "^1.0.0"
tiktoken = "^0.4.0"
unstructured = "^0.8.1"
psycopg2-binary = "^2.9.6"
langchain = {path="/Users/gordy/Downloads/langchain/libs/langchain/dist/langchain-0.0.240rc4.tar.gz"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
